+++ src/problems/problem_0006.f90
@@ -4,8 +4,8 @@ contains
 
 module subroutine euler0006(problem)
   type(problem_type), intent(inout) :: problem
-  integer(int32), parameter :: n = 100
-  integer(int32) :: i, sln
+  integer(int64), parameter :: n = 100
+  integer(int64) :: i, sln
 
   associate (array => [(i, i=1, n)])
     sln = sum(array)**2 - sum(array**2)
