+++ src/problems/problem_0009.f90
@@ -4,8 +4,8 @@ contains
 
 module subroutine euler0009(problem)
   type(problem_type), intent(inout) :: problem
-  integer(int32), parameter :: n = 1000
-  integer(int32) :: i, j, sln
+  integer(int64), parameter :: n = 1000
+  integer(int64) :: i, j, sln
 
   sln = 0
   do i = 1, n
