+++ src/problems/problem_0019.f90
@@ -4,7 +4,7 @@ contains
 
 module subroutine euler0019(problem)
   type(problem_type), intent(inout) :: problem
-  integer(int32) :: num_days(12), year, month, day_of_week, sln
+  integer(int64) :: num_days(12), year, month, day_of_week, sln
 
   num_days = 30 + [1, -2, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1]
   sln = 0
@@ -21,7 +21,7 @@ module subroutine euler0019(problem)
 end subroutine euler0019
 
 elemental logical function is_leap(year)
-  integer(int32), intent(in) :: year
+  integer(int64), intent(in) :: year
 
   is_leap = (mod(year, 4) == 0 .and. mod(year, 100) /= 0) .or. mod(year, 400) == 0
 end function is_leap
