+++ src/problems/problem_0033.f90
@@ -4,7 +4,7 @@ contains
 
 module subroutine euler0033(problem)
   type(problem_type), intent(inout) :: problem
-  integer(int32) :: denom, numer, counter, denom_prod, numer_prod
+  integer(int64) :: denom, numer, counter, denom_prod, numer_prod
 
   denom = 1
   numer = 1
